Sysex8 Handling improved. Now handles 8bit bytes and extra checks on lengths - fix type casts - #20 fixed

+++ src/umpProcessor.cpp
@@ -116,7 +116,7 @@ void umpProcessor::processUMP(uint32_t UMP){
                 mess.umpGroup = group;
                 mess.messageType = mt;
                 mess.form = (umpMess[0] >> 20) & 0xF;
-                mess.dataLength  = std::min((umpMess[0] >> 16) & 0xF, 6);
+                mess.dataLength  = (uint8_t)std::min((umpMess[0] >> 16) & 0xF, 6U);
                 uint8_t sysex[6];
 
                 if(mess.dataLength > 0)sysex[0] =  (umpMess[0] >> 8) & 0x7F;
@@ -366,7 +366,7 @@ void umpProcessor::processUMP(uint32_t UMP){
                 mess.messageType = mt;
                 mess.streamId  = (umpMess[0] >> 8) & 0xFF;
                 mess.form = status;
-                mess.dataLength  = std::min((umpMess[0] >> 16) & 0xF, 13);
+                mess.dataLength  = (uint8_t)std::min((umpMess[0] >> 16) & 0xF, 13U);
                 uint8_t sysex[13];
 
                 if(mess.dataLength >= 1)sysex[0] =  umpMess[0] & 0xFF;
